server/requirements.txt
Flask==3.0.0
pytest==7.4.3
pytest-cov==4.1.0
yt-dlp==2025.11.12
pydub==0.25.1
python-dotenv==1.0.0
azure-cognitiveservices-speech
flask-cors
audioop-lts  # Required for Python 3.13+ compatibility with pydub